- wisdom time  - cloudforms start vm

+++ monscale/rules.py
@@ -5,7 +5,7 @@ from django.conf import settings
 import logging, datetime
 from pytz import timezone
 from django.conf import settings
-from monscale.models import AlarmIndicator
+from monscale.models import AlarmIndicator, ActionIndicator
 
 
 def set_indicator(ctxt):
@@ -29,12 +29,42 @@ def set_indicator(ctxt):
         if time_diff.total_seconds() < float(threshold.time_limit):
             logging.debug("[set_indicator] exiting without triggering actions, reason: %s state alarm not enough time" % threshold)
             return
+    #checking wisdom time is passed
+    logging.debug("[set_indicator] checking wisdom time before launching any action")
     
+    try:
+        aindicator = service.action_indicators.all()[0]
+    
+        logging.debug("[set_indicator] ::: aindicator in DB")
+        """
+        If alarm indicator was found in DB means actions where trigger by its timestamp
+        thus need to check wisdom time against the alarm indicator timestamp
+        """
+        logging.debug("[set_indicator] Last action indicator: %s" % aindicator.timestamp)
+        time_diff = now - aindicator.timestamp
+        logging.debug("[set_indicator] Time diff: %s (%ss)" % (time_diff, time_diff.total_seconds()))
+        if time_diff.total_seconds() < float(service.wisdom_time):
+            logging.debug("[set_indicator] Wisdom time not over, not launching actions")
+            """
+            deleting threshold indicator thus, next indicator will start after the wisdom
+            """
+            indicator.delete()
+            return
+        """
+        delete the action indicator because wisdom time is over, actions must be launched
+        """
+        aindicator.delete()
+    except: pass
+
+    #lets launch the actions
     for action in service.action.all(): 
         action.to_redis(str(service))
         logging.debug("[set_indicator] Action: %s is queued" % action)
-        indicator.delete()
-    logging.debug("[set_indicator] exiting")
+    indicator.delete()
+    ActionIndicator(service=service).save()
+
+
+    
     
 def evaluate():
     mappings = get_mappings()
@@ -57,9 +87,9 @@ def evaluate():
             table.evaluate(ctxt)
             logging.info("[evaluate] service: %s has evaluate ALARM" % service)
             set_indicator(ctxt)
-        except Exception:
+        except Exception, er:
             try:
-                logging.debug("[evaluate] threshold has evaluate no alarm")
+                logging.debug("[evaluate] threshold has evaluate no alarm: %s" % er)
                 service.alarm_indicators.get().delete()
             except AlarmIndicator.DoesNotExist: pass
             print("No ha cumplido para el servicio: %s" % service)
